refactor: rename import for analyseBigFiles and remove legacy processBigFiles code
